Construct body-joint system

Construct body-joint system

User-oriented set up

ConfigDataType

Code-oriented construction

Before running any dynamics, we need to construct the body-joint system, connecting bodies by joints in the correct hierarchy using supplied configuration information. In order to do that, three key functions are used:

AddBody(id::Int, cb::ConfigBody)

AddJoint(id::Int, cj::ConfigJoint)

AssembleSystem(bs, js, sys)

which is responsible for add a body to the body system, add a joint to the joint system and assemble the body-joint system and fill in extra hierarchy information. These three functions are also bundled in the correct sequential order in

BuildChain(cbs, cjs, csys)

In order to solve for the rigid body-joint chain system, we choose body velocity $v$ and joint displacement $qJ$ as main variables in the time marching scheme. We also need to construct a structure that stores and updates the body-joint chain's intermediate variables such as all kinds of transformation matrices $X$, body position in the inertial space $x_i$ and so on. All information of the body-joint system is bundled into a BodyDyn structure.

Methods

Dyn3d.ConfigDataType.ConfigBody — Method.

the final plotting direction is: ^(y) |_____>(x) (-z) the coordinate for verts in 2d input is in [z,x]. So y direction only allow zero-width body

source

Dyn3d.ConstructSystem.Soln — Type.

This module construct the body-joint system by: 1. AddBody 2. AddJoint 3. AssembleSystem

source

Dyn3d.ConstructSystem.AddBody — Method.

AddBody(id::Int, cf::ConfigBody)

Using ConfigBody information, add one body to the body system bs

source

Dyn3d.ConstructSystem.AddJoint — Method.

AddJoint(id::Int, cf::ConfigJoint)

Using ConfigJoint information, add one joint to the joint system js

source

Dyn3d.ConstructSystem.AssembleSystem — Method.

AssembleSystem(bs::Vector{SingleBody}, js::Vector{SingleJoint}, sys::System)

With body system bs and joint system js, fill in extra hierarchy information about the whole system, connects bodies and joints.

source

Dyn3d.ConstructSystem.BuildChain — Method.

BuildChain(cbs::Vector{ConfigBody}, cjs::Vector{ConfigJoint}, csys::ConfigSystem)

Put AddBody, AddJoint and AssembleSystem in sequential order in a single function

source

Index

Dyn3d.ConfigDataType.ConfigBody

Dyn3d.ConstructSystem.Soln

Dyn3d.ConstructSystem.AddBody

Dyn3d.ConstructSystem.AddJoint

Dyn3d.ConstructSystem.AssembleSystem

Dyn3d.ConstructSystem.BuildChain